Refactoring and implementing package.go
The file package.go does a lot to make the rest of the implementation of
the functionality very simple. It is also much more reliable, as the
packages are properly inspected, instead of just relying on regular
expressions as I did in the past. Further, the actions have been
consolidated into the actions.go file.