API change: ~ReadAUR, +ReadAllAUR, -ConcurrentlyReadAUR
In order to stay relevant, the API has been changed to match the
change from AURv3 to AURv4.

ReadAUR signature and functionality has been changed.
ReadAllAUR replaces ConcurrentlyReadAUR as a more efficient
implementation.

Also, the AURPackage datatype has been made public. Statistics like
votes and popularity can be retrieved from this type. It is also
the type that is returned from ReadAUR and ReadAllAUR.

The Read*AUR functions also return the new NotFoundError in cases
where the package was not found.

Note: this commit has not been tested yet. If there are any bugs,
they will be fixed and commits will follow.