Improve speed of reading a directory of packages
The following function has been renamed:

    pacman.{ReadDir -> ReadEveryFileInDir}

This long name makes it clear that you probably don't want to use
it. The previous implementation of ReadDir has been modified to
require the filepath of the database, which it will use to get
metadata for packages it knows about. Any packages that are newer
than the database or that the database does not have are loaded
anyway.

This makes the following repoctl commands radically faster
(by at least a factor of 10):

    status
    update
    list
    down -u