Refactor repoctl to a simpler layout
In particular, everything that was previously the repoctl module
is now repoctl/repo, and the entire command line app, that was
previously in repoctl/cmd/repoctl is now in the root directory
of the repository.

This means it is no longer necessary to use go install ./...
with the ellipsis to install repoctl.